♻️ refactor: extract DatabaseManager migrations to sibling file
Move registerMigrations + per-version helpers (v1–v8) into
DatabaseManager+Migrations.swift to keep the main file under SwiftLint's
400-line file_length cap. Migrations are append-only and grow unboundedly;
this gives new versions room without touching the coordinator file. Pure
move — no behavior change. registerMigrations widens private → internal for
the cross-file call from makeMigrator(); the extension is `nonisolated`
(Data-layer type, swift-isolation Pattern 3).

See #748.
